backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
dotenv
alembic
dirtyjson
python-jose
pydantic[email]
passlib==1.7.4
bcrypt==4.0.1